Add lecture notes after WS2019 lecture
This is the overworked version of the lecture notes after I held
the lecture for the first time. The original lecture's content has
been reduced as I moved some of it to another lecture. Other
content will be reintroduced before I'll do the same lecture again
in 2020.

I also restructured this quite a bit, hopefully the new TOC makes
more sense. Dear students, be warned that my presentation slides
and this lecture notes are highly out-of-sync by now.

Signed-off-by: Andreas Happe <andreashappe@snikt.net>